+++ DatadogInternal/Tests/NetworkInstrumentation/NetworkInstrumentationFeatureTests.swift
@@ -24,22 +24,24 @@ class NetworkInstrumentationFeatureTests: XCTestCase {
     }
 
     override func tearDown() {
+        core?.get(feature: NetworkInstrumentationFeature.self)?.unbindAll()
         core = nil
         super.tearDown()
     }
 
     // MARK: - Interception Flow
 
-    func testGivenURLSessionWithDatadogDelegate_whenUsingTaskWithURL_itNotifiesInterceptor() {
-        let notifyInterceptionStart = expectation(description: "Notify interception did start")
-        let notifyInterceptionComplete = expectation(description: "Notify intercepion did complete")
+    func testGivenURLSessionWithDatadogDelegate_whenUsingTaskWithURL_itNotifiesInterceptor() throws {
+        let notifyInterceptionDidStart = expectation(description: "Notify interception did start")
+        let notifyInterceptionDidComplete = expectation(description: "Notify intercepion did complete")
         let server = ServerMock(delivery: .success(response: .mockResponseWith(statusCode: 200), data: .mock(ofSize: 10)))
 
-        handler.onInterceptionStart = { _ in notifyInterceptionStart.fulfill() }
-        handler.onInterceptionComplete = { _ in notifyInterceptionComplete.fulfill() }
+        handler.onInterceptionDidStart = { _ in notifyInterceptionDidStart.fulfill() }
+        handler.onInterceptionDidComplete = { _ in notifyInterceptionDidComplete.fulfill() }
 
         // Given
-        let delegate = DatadogURLSessionDelegate(in: core)
+        let delegate = MockDelegate()
+        try URLSessionInstrumentation.enableOrThrow(with: .init(delegateClass: MockDelegate.self), in: core)
         let session = server.getInterceptedURLSession(delegate: delegate)
 
         // When
@@ -49,31 +51,32 @@ class NetworkInstrumentationFeatureTests: XCTestCase {
         // Then
         wait(
             for: [
-                notifyInterceptionStart,
-                notifyInterceptionComplete
+                notifyInterceptionDidStart,
+                notifyInterceptionDidComplete
             ],
-            timeout: 0.5,
+            timeout: 1,
             enforceOrder: true
         )
         _ = server.waitAndReturnRequests(count: 1)
     }
 
-    func testGivenURLSessionWithDatadogDelegate_whenUsingTaskWithURLRequest_itNotifiesInterceptor() {
+    func testGivenURLSessionWithDatadogDelegate_whenUsingTaskWithURLRequest_itNotifiesInterceptor() throws {
         let notifyRequestMutation = expectation(description: "Notify request mutation")
-        let notifyInterceptionStart = expectation(description: "Notify interception did start")
-        let notifyInterceptionComplete = expectation(description: "Notify intercepion did complete")
+        let notifyInterceptionDidStart = expectation(description: "Notify interception did start")
+        let notifyInterceptionDidComplete = expectation(description: "Notify intercepion did complete")
         let server = ServerMock(delivery: .success(response: .mockResponseWith(statusCode: 200), data: .mock(ofSize: 10)))
 
         handler.onRequestMutation = { _, _ in notifyRequestMutation.fulfill() }
-        handler.onInterceptionStart = { _ in notifyInterceptionStart.fulfill() }
-        handler.onInterceptionComplete = { _ in notifyInterceptionComplete.fulfill() }
+        handler.onInterceptionDidStart = { _ in notifyInterceptionDidStart.fulfill() }
+        handler.onInterceptionDidComplete = { _ in notifyInterceptionDidComplete.fulfill() }
 
         // Given
         let url: URL = .mockAny()
         handler.firstPartyHosts = .init(
             hostsWithTracingHeaderTypes: [url.host!: [.datadog]]
         )
-        let delegate = DatadogURLSessionDelegate(in: core)
+        let delegate = MockDelegate()
+        try URLSessionInstrumentation.enableOrThrow(with: .init(delegateClass: MockDelegate.self), in: core)
         let session = server.getInterceptedURLSession(delegate: delegate)
 
         // When
@@ -85,10 +88,77 @@ class NetworkInstrumentationFeatureTests: XCTestCase {
         wait(
             for: [
                 notifyRequestMutation,
-                notifyInterceptionStart,
-                notifyInterceptionComplete
+                notifyInterceptionDidStart,
+                notifyInterceptionDidComplete
             ],
-            timeout: 0.5,
+            timeout: 1,
+            enforceOrder: true
+        )
+        _ = server.waitAndReturnRequests(count: 1)
+    }
+
+    @available(iOS 15.0, tvOS 15.0, *)
+    @MainActor
+    func testGivenURLSessionWithCustomDelegate_whenUsingAsyncDataFromURL_itNotifiesInterceptor() async throws {
+        let notifyInterceptionDidStart = expectation(description: "Notify interception did start")
+        let notifyInterceptionDidComplete = expectation(description: "Notify intercepion did complete")
+        let server = ServerMock(delivery: .success(response: .mockResponseWith(statusCode: 200), data: .mock(ofSize: 10)))
+
+        handler.onInterceptionDidStart = { _ in notifyInterceptionDidStart.fulfill() }
+        handler.onInterceptionDidComplete = { _ in notifyInterceptionDidComplete.fulfill() }
+
+        // Given
+        let delegate = MockDelegate()
+        try URLSessionInstrumentation.enableOrThrow(with: .init(delegateClass: MockDelegate.self), in: core)
+        let session = server.getInterceptedURLSession(delegate: delegate)
+
+        // When
+        _ = try await session.data(from: URL.mockAny(), delegate: delegate)
+
+        // Then
+        await fulfillment(
+            of: [
+                notifyInterceptionDidStart,
+                notifyInterceptionDidComplete
+            ],
+            timeout: 1,
+            enforceOrder: true
+        )
+        _ = server.waitAndReturnRequests(count: 1)
+    }
+
+    @available(iOS 15.0, tvOS 15.0, *)
+    @MainActor
+    func testGivenURLSessionWithCustomDelegate_whenUsingAsyncDataForURLRequest_itNotifiesInterceptor() async throws {
+        let notifyInterceptionDidStart = expectation(description: "Notify interception did start")
+        let notifyInterceptionDidComplete = expectation(description: "Notify intercepion did complete")
+        let server = ServerMock(delivery: .success(response: .mockResponseWith(statusCode: 200), data: .mock(ofSize: 10)))
+
+        handler.onInterceptionDidStart = { interception in
+            XCTAssertTrue(interception.isFirstPartyRequest)
+            notifyInterceptionDidStart.fulfill()
+        }
+        handler.onInterceptionDidComplete = { _ in notifyInterceptionDidComplete.fulfill() }
+
+        // Given
+        let url: URL = .mockAny()
+        handler.firstPartyHosts = .init(
+            hostsWithTracingHeaderTypes: [url.host!: [.datadog]]
+        )
+        let delegate = MockDelegate()
+        try URLSessionInstrumentation.enableOrThrow(with: .init(delegateClass: MockDelegate.self), in: core)
+        let session = server.getInterceptedURLSession(delegate: delegate)
+
+        // When
+        _ = try await session.data(for: URLRequest(url: url), delegate: delegate)
+
+        // Then
+        await fulfillment(
+            of: [
+                notifyInterceptionDidStart,
+                notifyInterceptionDidComplete
+            ],
+            timeout: 1,
             enforceOrder: true
         )
         _ = server.waitAndReturnRequests(count: 1)
@@ -97,21 +167,22 @@ class NetworkInstrumentationFeatureTests: XCTestCase {
     // MARK: - Interception Values
 
     func testGivenURLSessionWithDatadogDelegate_whenTaskCompletesWithFailure_itPassesAllValuesToTheInterceptor() throws {
-        let notifyInterceptionStart = expectation(description: "Notify interception did start")
-        let notifyInterceptionComplete = expectation(description: "Notify intercepion did complete")
-        notifyInterceptionStart.expectedFulfillmentCount = 2
-        notifyInterceptionComplete.expectedFulfillmentCount = 2
+        let notifyInterceptionDidStart = expectation(description: "Notify interception did start")
+        let notifyInterceptionDidComplete = expectation(description: "Notify intercepion did complete")
+        notifyInterceptionDidStart.expectedFulfillmentCount = 2
+        notifyInterceptionDidComplete.expectedFulfillmentCount = 2
 
         let expectedError = NSError(domain: "network", code: 999, userInfo: [NSLocalizedDescriptionKey: "some error"])
         let server = ServerMock(delivery: .failure(error: expectedError))
 
-        handler.onInterceptionStart = { _ in notifyInterceptionStart.fulfill() }
-        handler.onInterceptionComplete = { _ in notifyInterceptionComplete.fulfill() }
+        handler.onInterceptionDidStart = { _ in notifyInterceptionDidStart.fulfill() }
+        handler.onInterceptionDidComplete = { _ in notifyInterceptionDidComplete.fulfill() }
 
         let dateBeforeAnyRequests = Date()
 
         // Given
-        let delegate = DatadogURLSessionDelegate(in: core)
+        let delegate = MockDelegate()
+        try URLSessionInstrumentation.enableOrThrow(with: .init(delegateClass: MockDelegate.self), in: core)
         let session = server.getInterceptedURLSession(delegate: delegate)
 
         // When
@@ -144,21 +215,22 @@ class NetworkInstrumentationFeatureTests: XCTestCase {
     }
 
     func testGivenURLSessionWithDatadogDelegate_whenTaskCompletesWithSuccess_itPassesAllValuesToTheInterceptor() throws {
-        let notifyInterceptionStart = expectation(description: "Notify interception did start")
-        let notifyInterceptionComplete = expectation(description: "Notify intercepion did complete")
-        notifyInterceptionStart.expectedFulfillmentCount = 2
-        notifyInterceptionComplete.expectedFulfillmentCount = 2
+        let notifyInterceptionDidStart = expectation(description: "Notify interception did start")
+        let notifyInterceptionDidComplete = expectation(description: "Notify intercepion did complete")
+        notifyInterceptionDidStart.expectedFulfillmentCount = 2
+        notifyInterceptionDidComplete.expectedFulfillmentCount = 2
 
         let randomData: Data = .mockRandom()
         let server = ServerMock(delivery: .success(response: .mockResponseWith(statusCode: 200), data: randomData))
 
-        handler.onInterceptionStart = { _ in notifyInterceptionStart.fulfill() }
-        handler.onInterceptionComplete = { _ in notifyInterceptionComplete.fulfill() }
+        handler.onInterceptionDidStart = { _ in notifyInterceptionDidStart.fulfill() }
+        handler.onInterceptionDidComplete = { _ in notifyInterceptionDidComplete.fulfill() }
 
         let dateBeforeAnyRequests = Date()
 
         // Given
-        let delegate = DatadogURLSessionDelegate(in: core)
+        let delegate = MockDelegate()
+        try URLSessionInstrumentation.enableOrThrow(with: .init(delegateClass: MockDelegate.self), in: core)
         let session = server.getInterceptedURLSession(delegate: delegate)
 
         // When
@@ -173,7 +245,7 @@ class NetworkInstrumentationFeatureTests: XCTestCase {
             .resume()
 
         // Then
-        waitForExpectations(timeout: 0.5, handler: nil)
+        waitForExpectations(timeout: 1, handler: nil)
         _ = server.waitAndReturnRequests(count: 1)
 
         let dateAfterAllRequests = Date()
@@ -188,6 +260,57 @@ class NetworkInstrumentationFeatureTests: XCTestCase {
             XCTAssertEqual(interception.data, randomData)
             XCTAssertNotNil(interception.completion)
             XCTAssertNil(interception.completion?.error)
+        }
+    }
+
+    @available(iOS 15.0, tvOS 15.0, *)
+    @MainActor
+    func testGivenURLSessionWithCustomDelegate_whenUsingAsyncData_itPassesAllValuesToTheInterceptor() async throws {
+        let notifyInterceptionDidStart = expectation(description: "Notify interception did start")
+        let notifyInterceptionDidComplete = expectation(description: "Notify intercepion did complete")
+        notifyInterceptionDidStart.expectedFulfillmentCount = 2
+        notifyInterceptionDidComplete.expectedFulfillmentCount = 2
+
+        let expectedError = NSError(domain: "network", code: 999, userInfo: [NSLocalizedDescriptionKey: "some error"])
+        let server = ServerMock(delivery: .failure(error: expectedError))
+
+        handler.onInterceptionDidStart = { _ in notifyInterceptionDidStart.fulfill() }
+        handler.onInterceptionDidComplete = { _ in notifyInterceptionDidComplete.fulfill() }
+
+        let dateBeforeAnyRequests = Date()
+
+        // Given
+        let delegate = MockDelegate()
+        try URLSessionInstrumentation.enableOrThrow(with: .init(delegateClass: MockDelegate.self), in: core)
+        let session = server.getInterceptedURLSession(delegate: delegate)
+
+        // When
+        let url1: URL = .mockRandom()
+        _ = try? await session.data(from: url1, delegate: delegate)
+
+        let url2: URL = .mockRandom()
+        _ = try? await session.data(for: URLRequest(url: url2), delegate: delegate)
+
+        // Then
+        await fulfillment(
+            of: [
+                notifyInterceptionDidStart,
+                notifyInterceptionDidComplete
+            ],
+            timeout: 1,
+            enforceOrder: true
+        )
+        _ = server.waitAndReturnRequests(count: 2)
+
+        let dateAfterAllRequests = Date()
+
+        XCTAssertEqual(handler.interceptions.count, 2, "Interceptor should record metrics for 2 tasks")
+
+        handler.interceptions.forEach { id, interception in
+            XCTAssertGreaterThan(interception.metrics?.fetch.start ?? .distantPast, dateBeforeAnyRequests)
+            XCTAssertLessThan(interception.metrics?.fetch.end ?? .distantFuture, dateAfterAllRequests)
+            XCTAssertNil(interception.data, "Data should not be recorded for \(id)")
+            XCTAssertEqual((interception.completion?.error as? NSError)?.localizedDescription, "some error")
         }
     }
 
@@ -328,21 +451,21 @@ class NetworkInstrumentationFeatureTests: XCTestCase {
     // MARK: - First Party Hosts
 
     func testGivenHandler_whenInterceptingRequests_itDetectFirstPartyHost() throws {
-        let notifyInterceptionStart = expectation(description: "Notify interception did start")
+        let notifyInterceptionDidStart = expectation(description: "Notify interception did start")
         let server = ServerMock(delivery: .success(response: .mockResponseWith(statusCode: 200), data: .mock(ofSize: 10)))
 
         // Given
-        let delegate = DatadogURLSessionDelegate(
-            in: core,
-            additionalFirstPartyHostsWithHeaderTypes: ["test.com": [.datadog]]
-        )
+        let delegate = MockDelegate()
+        let firstPartyHosts: URLSessionInstrumentation.FirstPartyHostsTracing = .traceWithHeaders(hostsWithHeaders: ["test.com": [.datadog]])
+        try URLSessionInstrumentation.enableOrThrow(with: .init(delegateClass: MockDelegate.self, firstPartyHostsTracing: firstPartyHosts), in: core)
+
         let session = server.getInterceptedURLSession(delegate: delegate)
         let request: URLRequest = .mockWith(url: "https://test.com")
 
-        handler.onInterceptionStart = {
+        handler.onInterceptionDidStart = {
             // Then
             XCTAssertTrue($0.isFirstPartyRequest)
-            notifyInterceptionStart.fulfill()
+            notifyInterceptionDidStart.fulfill()
         }
 
         // When
@@ -356,24 +479,24 @@ class NetworkInstrumentationFeatureTests: XCTestCase {
     }
 
     func testGivenDelegateSubclass_whenInterceptingRequests_itDetectFirstPartyHost() throws {
-        let notifyInterceptionStart = expectation(description: "Notify interception did start")
-        handler.onInterceptionStart = { _ in notifyInterceptionStart.fulfill() }
+        let notifyInterceptionDidStart = expectation(description: "Notify interception did start")
+        handler.onInterceptionDidStart = { _ in notifyInterceptionDidStart.fulfill() }
         let server = ServerMock(delivery: .success(response: .mockResponseWith(statusCode: 200), data: .mock(ofSize: 10)))
 
         // Given
-        class SubclassDelegate: DatadogURLSessionDelegate {}
 
-        let delegate = SubclassDelegate(
+        let delegate = DatadogURLSessionDelegate(
             in: core,
             additionalFirstPartyHostsWithHeaderTypes: ["test.com": [.datadog]]
         )
+
         let session = server.getInterceptedURLSession(delegate: delegate)
         let request: URLRequest = .mockWith(url: "https://test.com")
 
-        handler.onInterceptionStart = {
+        handler.onInterceptionDidStart = {
             // Then
             XCTAssertTrue($0.isFirstPartyRequest)
-            notifyInterceptionStart.fulfill()
+            notifyInterceptionDidStart.fulfill()
         }
 
         // When
@@ -387,8 +510,8 @@ class NetworkInstrumentationFeatureTests: XCTestCase {
     }
 
     func testGivenCompositeDelegate_whenInterceptingRequests_itDetectFirstPartyHost() throws {
-        let notifyInterceptionStart = expectation(description: "Notify interception did start")
-        handler.onInterceptionStart = { _ in notifyInterceptionStart.fulfill() }
+        let notifyInterceptionDidStart = expectation(description: "Notify interception did start")
+        handler.onInterceptionDidStart = { _ in notifyInterceptionDidStart.fulfill() }
         let server = ServerMock(delivery: .success(response: .mockResponseWith(statusCode: 200), data: .mock(ofSize: 10)))
 
         // Given
@@ -409,10 +532,10 @@ class NetworkInstrumentationFeatureTests: XCTestCase {
         let session = server.getInterceptedURLSession(delegate: delegate)
         let request: URLRequest = .mockWith(url: "https://test.com")
 
-        handler.onInterceptionStart = {
+        handler.onInterceptionDidStart = {
             // Then
             XCTAssertTrue($0.isFirstPartyRequest)
-            notifyInterceptionStart.fulfill()
+            notifyInterceptionDidStart.fulfill()
         }
 
         // When
@@ -450,5 +573,8 @@ class NetworkInstrumentationFeatureTests: XCTestCase {
             iterations: 50
         )
         // swiftlint:enable opening_brace trailing_closure
+    }
+
+    class MockDelegate: NSObject, URLSessionDataDelegate {
     }
 }